Viridian test updates
